@@ -20,12 +20,12 @@
     "gh-pages": "^1.2.0",
     "history": "^4.10.1",
     "npm": "^7.5.3",
-    "react": "^16.14.0",
+    "react": "^18.2.0",
     "react-bootstrap": "^0.32.4",
     "react-copy-to-clipboard": "^5.0.3",
-    "react-dom": "^16.14.0",
-    "react-router": "^4.3.1",
-    "react-router-dom": "^4.3.1",
+    "react-dom": "^18.2.0",
+    "react-router": "^6.20.0",
+    "react-router-dom": "^6.20.0",
     "react-scripts": "5.0.1",
     "react-tooltip": "^5.21.6",
     "yarn": "^1.22.10"
@@ -34,9 +34,9 @@
     "@babel/eslint-parser": "^7.23.3",
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "@types/node": "^20.8.9",
-    "@types/react": "^18.2.33",
+    "@types/react": "^18.2.38",
     "@types/react-copy-to-clipboard": "^5.0.3",
-    "@types/react-dom": "^18.2.14",
+    "@types/react-dom": "^18.2.17",
     "@types/react-router": "^5.1.20",
     "@types/react-router-dom": "^5.3.3",
     "@typescript-eslint/eslint-plugin": "^6.11.0",
